refactor(pi-conductor): extract human gate UI
